split fixtures into smaller modules
