+++ test/origin_simulator/admin_router_test.exs
@@ -2,7 +2,7 @@ defmodule OriginSimulator.AdminRouterTest do
   use ExUnit.Case
   use Plug.Test
 
-  import Fixtures
+  import Fixtures.Recipes
   import TestHelpers
   import OriginSimulator, only: [recipe_not_set: 1]
