+++ test/test_helper.exs
@@ -1,4 +1,5 @@
 Code.require_file("fixtures/fixtures.exs", __DIR__)
+Code.require_file("fixtures/recipes.exs", __DIR__)
 
 ExUnit.start()
